requirements.txt
aiohttp==3.9.5
beautifulsoup4==4.12.3
python-dotenv==1.0.1
aiogram==3.4.1
requests==2.31.0
lxml==4.9.3
urllib3==2.1.0
html5lib==1.1 